# api-docs.json
{"openapi":"3.0.1","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"paths":{"/venues/create":{"post":{"tags":["venues-controller"],"operationId":"createVenue","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/VenuesDTO"}}},"required":true},"responses":{"200":{"description":"OK"}}}},"/users/set/users":{"post":{"tags":["users-controller"],"operationId":"SetUsers","requestBody":{"content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/UsersDto"}}}},"required":true},"responses":{"200":{"description":"OK"}}}},"/types/create":{"post":{"tags":["types-controller"],"operationId":"createType","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TypesDTO"}}},"required":true},"responses":{"200":{"description":"OK"}}}},"/events/create":{"post":{"tags":["events-controller"],"operationId":"createEvent","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/EventsDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/contacts/create":{"post":{"tags":["contacts-controller"],"operationId":"setContacts","parameters":[{"name":"contactsDTO","in":"query","required":true,"schema":{"$ref":"#/components/schemas/ContactsDTO"}}],"responses":{"200":{"description":"OK"}}}},"/clubs/create":{"post":{"tags":["clubs-controller"],"operationId":"setClubs","parameters":[{"name":"clubsDTO","in":"query","required":true,"schema":{"$ref":"#/components/schemas/ClubsDTO"}}],"responses":{"200":{"description":"OK"}}}},"/venues/all":{"get":{"tags":["venues-controller"],"operationId":"getAllVenues","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/VenuesDTO"}}}}}}}},"/users/get/users":{"get":{"tags":["users-controller"],"operationId":"getAllUsers","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Users"}}}}}}}},"/types/all":{"get":{"tags":["types-controller"],"operationId":"getAllTypes","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/TypesDTO"}}}}}}}},"/events/{id}/image":{"get":{"tags":["events-controller"],"operationId":"getEventImage","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"type":"string","format":"byte"}}}}}}}},"/events/":{"get":{"tags":["events-controller"],"operationId":"getAllEvents","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Events"}}}}}}}},"/contacts/all":{"get":{"tags":["contacts-controller"],"operationId":"getAllContacts","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Contacts"}}}}}}}},"/clubs/all":{"get":{"tags":["clubs-controller"],"operationId":"getAllClubs","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Clubs"}}}}}}}}},"components":{"schemas":{"VenuesDTO":{"type":"object","properties":{"venueId":{"type":"integer","format":"int64"},"venueName":{"type":"string"}}},"UsersDto":{"type":"object","properties":{"userID":{"type":"integer","format":"int64"},"username":{"type":"string"},"email":{"type":"string"},"password":{"type":"string"},"role":{"type":"string"}}},"TypesDTO":{"type":"object","properties":{"typeId":{"type":"integer","format":"int64"},"name":{"type":"string"},"email":{"type":"string"},"phone":{"type":"string"}}},"ContactsDTO":{"type":"object","properties":{"contactId":{"type":"integer","format":"int64"},"name":{"type":"string"},"email":{"type":"string"},"phone":{"type":"string"}}},"EventsDTO":{"type":"object","properties":{"eventId":{"type":"integer","format":"int64"},"name":{"type":"string"},"date":{"type":"string"},"startTime":{"type":"string"},"endTime":{"type":"string"},"venue":{"$ref":"#/components/schemas/VenuesDTO"},"description":{"type":"string"},"type":{"$ref":"#/components/schemas/TypesDTO"},"contact":{"$ref":"#/components/schemas/ContactsDTO"},"clubId":{"type":"integer","format":"int64"},"image":{"type":"string"}}},"ClubsDTO":{"type":"object","properties":{"clubId":{"type":"integer","format":"int64"},"clubName":{"type":"string"},"clubDescription":{"type":"string"},"clubContact":{"type":"string"}}},"Users":{"type":"object","properties":{"userID":{"type":"integer","format":"int64"},"username":{"type":"string"},"email":{"type":"string"},"password":{"type":"string"},"role":{"type":"string"}}},"Clubs":{"type":"object","properties":{"clubId":{"type":"integer","format":"int64"},"clubName":{"type":"string"},"clubDescription":{"type":"string"},"clubContact":{"type":"string"}}},"Contacts":{"type":"object","properties":{"contactId":{"type":"integer","format":"int64"},"name":{"type":"string"},"email":{"type":"string"},"phone":{"type":"string"}}},"Events":{"type":"object","properties":{"eventId":{"type":"integer","format":"int64"},"name":{"type":"string"},"date":{"type":"string"},"startTime":{"type":"string"},"endTime":{"type":"string"},"venue":{"$ref":"#/components/schemas/Venues"},"description":{"type":"string"},"type":{"$ref":"#/components/schemas/Types"},"contact":{"$ref":"#/components/schemas/Contacts"},"club":{"$ref":"#/components/schemas/Clubs"},"image":{"type":"array","items":{"type":"string","format":"byte"}}}},"Types":{"type":"object","properties":{"contactId":{"type":"integer","format":"int64"},"name":{"type":"string"},"email":{"type":"string"},"phone":{"type":"string"}}},"Venues":{"type":"object","properties":{"venueId":{"type":"integer","format":"int64"},"venueName":{"type":"string"}}}}}}