# car-backend-v/src/main/java/com/example/cap06backend/cap06backend/Contacts/Entities/Contacts.java
package com.example.cap06backend.cap06backend.Contacts.Entities;

import jakarta.persistence.*;
import lombok.Getter;
import lombok.Setter;
import org.springframework.stereotype.Service;

@Getter
@Setter
@Entity
@Table(name = "contacts")
public class Contacts{
    @Id
    @GeneratedValue(strategy = GenerationType.IDENTITY)
    @Column(name = "ContactID")
    private Long contactId;

    @Column(name = "Name")
    private String name;

    @Column(name = "Email")
    private String email;

    @Column(name = "Phone")
    private String phone;
}