# car-backend-v/src/main/java/com/example/cap06backend/cap06backend/Users/Entities/Users.java
package com.example.cap06backend.cap06backend.Users.Entities;

import jakarta.persistence.*;
import lombok.Getter;
import lombok.Setter;

@Entity
@Getter
@Setter
@Table(name = "users")
public class Users {
    @Id
    @GeneratedValue(strategy = GenerationType.IDENTITY)
    private Long userID;

    private String username;
    private String email;
    private String password;
    private String role;

}